improved permission listeners (repeated code)
